+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/artefactsummary/ArtefactSummaryService.java
@@ -1,19 +1,10 @@
 package uk.gov.hmcts.reform.pip.publication.services.service.artefactsummary;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.pip.publication.services.models.external.ListType;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Locale;
 
 
 @Service
@@ -29,12 +20,11 @@ public class ArtefactSummaryService {
     @Autowired
     SjpPublicList sjpPublicList;
 
-    private static final String COURT_LISTS = "courtLists";
-    private static final String COURT_HOUSE = "courtHouse";
-    private static final String COURT_ROOM = "courtRoom";
-    private static final String SESSION = "session";
-    private static final String SITTINGS = "sittings";
-    private static final String INDIVIDUAL_DETAILS = "individualDetails";
+    @Autowired
+    SjpPressList sjpPressList;
+
+    @Autowired
+    CivilDailyCauseList civilDailyCauseList;
 
     /**
      * Parent class to route based on list types.
@@ -50,253 +40,13 @@ public class ArtefactSummaryService {
             case SJP_PUBLIC_LIST:
                 return sjpPublicList.artefactSummarysjpPublic(payload);
             case SJP_PRESS_LIST:
-                return artefactSummarysjpPress(payload);
+                return sjpPressList.artefactSummarysjpPress(payload);
             case CIVIL_DAILY_CAUSE_LIST:
-                return artefactSummaryCivilDailyCause(payload);
+                return civilDailyCauseList.artefactSummaryCivilDailyCause(payload);
             case FAMILY_DAILY_CAUSE_LIST:
                 return "fam_cause";
             default:
                 return "";
         }
     }
-
-    /**
-     * Civil cause list parent method - iterates on courtHouse/courtList - if these need to be shown in further
-     * iterations, do it here.
-     *
-     * @param payload - json body.
-     * @return - string for output.
-     * @throws JsonProcessingException - jackson req.
-     */
-    private String artefactSummaryCivilDailyCause(String payload) throws JsonProcessingException {
-        StringBuilder output = new StringBuilder("");
-        JsonNode node = new ObjectMapper().readTree(payload);
-        Iterator<JsonNode> courtHouseNode = node.get(COURT_LISTS).elements();
-        while (courtHouseNode.hasNext()) {
-            JsonNode thisCourtHouse = courtHouseNode.next().get(COURT_HOUSE);
-            output.append(processCivilDailyCourtRooms(thisCourtHouse)).append('\n');
-        }
-        return output.toString();
-    }
-
-    /**
-     * court room iteration - cycles through courtrooms and deals with routing for hearing channel, judiciary
-     * and sitting methods.
-     *
-     * @param node - jsonnode of courtrooms.
-     * @return string with above-mentioned info.
-     */
-    private String processCivilDailyCourtRooms(JsonNode node) {
-        Iterator<JsonNode> courtRoomNode = node.get(COURT_ROOM).elements();
-        ObjectMapper mapper = new ObjectMapper();
-        StringBuilder outputString = new StringBuilder();
-        List<String> sessionChannel;
-        TypeReference<List<String>> typeReference = new TypeReference<>() {
-        };
-        while (courtRoomNode.hasNext()) {
-            JsonNode thisCourtRoom = courtRoomNode.next();
-            JsonNode sessionChannelNode = thisCourtRoom.get(SESSION).get(0).path("sessionChannel");
-            sessionChannel = mapper.convertValue(sessionChannelNode, typeReference);
-            outputString.append("\n\nCourtroom: ").append(thisCourtRoom.get("courtRoomName").asText())
-            .append(processCivilDailySittings(thisCourtRoom, sessionChannel))
-            .append(processCivilDailyJudiciary(thisCourtRoom));
-        }
-        return outputString.toString();
-    }
-
-    /**
-     * Judiciary iteration - gets title and known as fields from judiciary node.
-     *
-     * @param node - node of judiciary.
-     * @return judiciary string
-     */
-    private String processCivilDailyJudiciary(JsonNode node) {
-        JsonNode judiciaryNode = node.get(SESSION).get(0).get("judiciary");
-        if (judiciaryNode.isEmpty()) {
-            return "";
-        }
-        Iterator<JsonNode> johNode = judiciaryNode.elements();
-        StringBuilder johName = new StringBuilder("\nJudiciary: ");
-        while (johNode.hasNext()) {
-            JsonNode currentJoh = johNode.next();
-            String title = currentJoh.path("johTitle").asText();
-            String knownAs = currentJoh.path("johKnownAs").asText();
-            johName.append(title).append(' ');
-            johName.append(knownAs);
-            if (johNode.hasNext()) {
-                johName.append(", ");
-            }
-        }
-        return johName.toString();
-    }
-
-    /**
-     * sitting iteration class - deals with hearing channel, start time and hearing data (e.g. case names, refs etc)
-     *
-     * @param node - node of sittings.
-     * @param sessionChannel - session channel passed in from parent method - overridden if sitting level channel
-     *                       exists.
-     * @return string of these bits.
-     */
-    private String processCivilDailySittings(JsonNode node, List<String> sessionChannel) {
-        JsonNode sittingNode = node.get(SESSION).get(0).get(SITTINGS);
-        Iterator<JsonNode> sittingIterator = sittingNode.elements();
-        StringBuilder outputString = new StringBuilder(26);
-        int counter = 1;
-        // below line is due to pmd "avoid using literals in conditional statements" rule.
-        boolean sittingNodeSizeBool = sittingNode.size() > 1;
-        while (sittingIterator.hasNext()) {
-            outputString.append("\n•Hearing");
-            if (sittingNodeSizeBool) {
-                outputString.append(' ').append(counter);
-                counter += 1;
-            }
-            JsonNode currentSitting = sittingIterator.next();
-            DateTimeFormatter inputfmt = DateTimeFormatter.ISO_DATE_TIME;
-            DateTimeFormatter outputfmt = DateTimeFormatter.ofPattern("hh:mm a");
-            String startTime =
-                outputfmt.format(LocalDateTime.from(inputfmt.parse(currentSitting.get("sittingStart").asText())));
-            outputString.append(processCivilDailyHearings(currentSitting))
-                .append(": \nStart Time: ").append(startTime)
-                .append(processCivilDailyChannels(sessionChannel, currentSitting.path("channel")));
-        }
-        return outputString.toString();
-    }
-
-    /**
-     * hearing channel handler. Sitting channel takes precedence over session channel if both exist (session channel
-     * is mandatory, however).
-     *
-     * @param sessionChannel - mentioned above.
-     * @param currentSittingNode - node for getting current sitting channel data.
-     * @return - string of correct channel.
-     */
-    private String processCivilDailyChannels(List<String> sessionChannel, JsonNode currentSittingNode) {
-        StringBuilder outputString = new StringBuilder();
-        ObjectMapper mapper = new ObjectMapper();
-        if (currentSittingNode.isMissingNode()) {
-            for (String channel : sessionChannel) {
-                outputString.append('\n').append(channel);
-            }
-        } else {
-            List<String> channelList = mapper.convertValue(currentSittingNode, new TypeReference<List<String>>() {
-            });
-            for (String channel : channelList) {
-                outputString.append('\n').append(channel);
-            }
-        }
-        return outputString.toString();
-    }
-
-    /**
-     * hearing iterator - gets case names, refs and hearing types.
-     *
-     * @param node - iterator of hearings.
-     * @return String with that stuff in it.
-     */
-    private String processCivilDailyHearings(JsonNode node) {
-        StringBuilder output = new StringBuilder(47);
-        Iterator<JsonNode> hearingNode = node.get("hearing").elements();
-        while (hearingNode.hasNext()) {
-            JsonNode currentHearing = hearingNode.next();
-            String hearingType = currentHearing.get("hearingType").asText();
-            String caseName = currentHearing.get("case").get(0).get("caseName").asText();
-            String caseNumber = currentHearing.get("case").get(0).get("caseNumber").asText();
-            output.append("\nCase Name: ").append(caseName)
-                .append("\nCase Reference: ").append(caseNumber)
-                .append("\nHearing Type: ").append(hearingType);
-        }
-        return output.toString();
-    }
-
-    /**
-     * sjp press parent method - iterates over session data. Routes to specific methods which handle offences and
-     * judiciary roles.
-     *
-     * @param payload - json body.
-     * @return String with final summary data.
-     * @throws JsonProcessingException - jackson req.
-     */
-    private String artefactSummarysjpPress(String payload) throws JsonProcessingException {
-        StringBuilder output = new StringBuilder();
-        JsonNode node = new ObjectMapper().readTree(payload);
-        Iterator<JsonNode> endNode =
-            node.get("courtLists").get(0).get("courtHouse").get("courtRoom").get(0).get(SESSION).get(0).get(
-                "sittings").get(0).get("hearing").elements();
-        while (endNode.hasNext()) {
-            output.append('•');
-            JsonNode currentCase = endNode.next();
-            output.append(processRolessjpPress(currentCase));
-            output.append(processOffencessjpPress(currentCase.get("offence"))).append('\n');
-        }
-        return output.toString();
-    }
-
-    /**
-     * offences iterator method - handles logic of accused of single or multiple offences and returns output string.
-     *
-     * @param offencesNode - iterator on offences.
-     * @return string with offence data.
-     */
-    private String processOffencessjpPress(JsonNode offencesNode) {
-        StringBuilder outputString = new StringBuilder();
-        // below line is due to pmd "avoid using literals in conditional statements" rule.
-        boolean offencesNodeSizeBool = offencesNode.size() > 1;
-        if (offencesNodeSizeBool) {
-            Iterator<JsonNode> offences = offencesNode.elements();
-            int counter = 1;
-            while (offences.hasNext()) {
-                JsonNode thisOffence = offences.next();
-                outputString.append("\nOffence ").append(counter).append(": ")
-                    .append(thisOffence.get("offenceTitle").asText());
-                outputString.append(processReportingRestrictionsjpPress(thisOffence));
-                counter += 1;
-            }
-        } else {
-            outputString.append("\nOffence: ").append(offencesNode.get(0).get("offenceTitle").asText())
-                .append(processReportingRestrictionsjpPress(offencesNode.get(0)));
-        }
-        return outputString.toString();
-    }
-
-    /**
-     * handles reporting restrictions for sjp press.
-     *
-     * @param node - node which is checked for reporting restriction.
-     * @return - text based on whether restriction exists.
-     */
-    private String processReportingRestrictionsjpPress(JsonNode node) {
-        boolean restriction = node.get("reportingRestriction").asBoolean();
-        if (restriction) {
-            return "(Reporting restriction)";
-        }
-        return "";
-    }
-
-    /**
-     * role iteration method for sjp press.
-     *
-     * @param party - iterator of party.
-     * @return list of roles.
-     */
-    private String processRolessjpPress(JsonNode party) {
-        Iterator<JsonNode> partyNode = party.get("party").elements();
-        String accused = "";
-        String postCode = "";
-        String prosecutor = "";
-        while (partyNode.hasNext()) {
-            JsonNode currentParty = partyNode.next();
-            if ("accused".equals(currentParty.get("partyRole").asText().toLowerCase(Locale.ROOT))) {
-                String forename = currentParty.get(INDIVIDUAL_DETAILS).get("individualForenames").asText();
-                String surname = currentParty.get(INDIVIDUAL_DETAILS).get("individualSurname").asText();
-                postCode = currentParty.get(INDIVIDUAL_DETAILS).get("address").get("postCode").asText();
-                accused = forename + " " + surname;
-            } else {
-                prosecutor = currentParty.get("organisationDetails").get("organisationName").asText();
-            }
-        }
-        return "Accused: " + accused + "\nPostcode: " + postCode + "\nProsecutor: " + prosecutor;
-    }
-
-
 }
